Important BUG FIX! Fixed a bug that would cause flows would become stalled.
If Qdisc drops a single packet, TCP would assume the entire segment is dropped.
In the worst case, a flow can become stalled because it sent data TCP does not
think has been sent yet.  In this case, incoming ACKs will be seen as invalid.
This was a huge pain in the ass to find.